Can now create listings, with error handling for cancel and creating without an image. Temporary fix for consoles without a GameCategory
Added "none" as a gameCategory in case the listing is a Console. May be subject to change. Will bring up this issue to the next meeting.

Cancel listing now prompts the user if they really want to cancel.

Listings are now displayed on the page when created. The image however isn't displayed yet.