Last push to branch | DEAD BRANCH (see comment)
We have to re-scaffold the project in main branch with the identity option selected to get a working login as no documentation mentions how to do this without it for .NET 8.

The scaffold can not be added to this version of the program due to an unfixed Visual Studio error that's been around since .NET 6.