Moved Models to Data-folder | ApplicationUser replaces Account
Moved the contents of the Models folder to Data (this follows microsofts "best-practice" apparently).

Removed Account class and added it's properties to ApplicationUser which already contained a lot of similar functionality. This also makes it easier for us when we want to Authorize a user to perform CRUD operations on their own listings.